change bin_length type

+++ cpp_src/EliasFano.h
@@ -99,7 +99,7 @@ class EliasFanoDB
   const py::tuple getCellTypeMatrix(const CellTypeName& cell_type, const py::list& gene_list) const;
   // Add this to the public section of the EliasFanoDB class
 
-  py::dict getCellTypeExpressionBinData(const std::string &cell_type, const py::list &gene_list, const int bin_length = 1) const;
+  py::dict getCellTypeExpressionBinData(const std::string &cell_type, const py::list &gene_list, const float bin_length = 1.0) const;
 
   int numberOfCellTypes(const py::list&) const;
